backend/pyproject.toml
[project]
name = "todo-app-backend"
version = "0.1.0"
description = "Self-hosted Todo Application Backend"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.0",
    "django-filter>=23.0",
    "python-decouple>=3.8",
    "pillow>=10.0",
    "dj-database-url>=2.1",
    "djangorestframework-simplejwt>=5.3",
    "django-extensions>=3.2",
]

[project.optional-dependencies]
dev = [
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "pytest>=7.4",
    "pytest-django>=4.5",
    "factory-boy>=3.3",
    "coverage>=7.3",
    "pre-commit>=3.4",
    "django-debug-toolbar>=4.2",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
skip = ["migrations"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/tests/*",
    "manage.py",
    "*/settings/*",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]